+++ src/PostHog.AI/CHANGELOG.md
@@ -1,5 +1,12 @@
 # PostHog.AI
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [293539c]
+  - PostHog@2.6.2
+
 ## 0.1.1
 
 ### Patch Changes